requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.22.0
pillow>=9.0.0
opencv-python>=4.7.0
pyyaml>=6.0

# YOLO implementations
ultralytics>=8.0.0
transformers>=4.30.0

# Data handling and augmentation
albumentations>=1.3.0
pycocotools>=2.0.6

# Training utilities
pytorch-lightning>=2.0.0
tensorboard>=2.12.0
wandb>=0.15.0

# Model export and deployment
onnx>=1.13.0
onnxruntime>=1.14.0

# Utilities
tqdm>=4.64.0
matplotlib>=3.7.0
pandas>=1.5.0

# Development
pytest>=7.3.1
black>=23.3.0
isort>=5.12.0
mypy>=1.3.0
